% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civ3}
\alias{civ3}
\title{Data on Civ 3}
\format{
A data frame with 31 rows and 5 columns.
\describe{
 \item{name}{Name of the civilisation}
 \item{leader}{The historical leader of the civilisation}
 \item{strengths}{The two strengths of this civilisation, separated by a comma}
 \item{starting_techs}{The two starting technologies of this civilisation, separated by a comma}
 \item{unique_unit}{The unique unit of this civilisation}
}
}
\source{
Obtained from the Civilization Wiki https://civilization.fandom.com/wiki/Civilizations_(Civ3)
}
\usage{
civ3
}
\description{
A dataset containing information about the civilisations in Sid Meier's Civilization 3
}
\examples{
data(civ3)
}
\keyword{datasets}
